updated firebase upload and pdf creation
