+++ modules/onboarding/src/main/kotlin/edu/stanford/spezi/module/onboarding/consent/PdfService.kt
@@ -2,6 +2,6 @@ package edu.stanford.spezi.module.onboarding.consent
 
 interface PdfService {
 
-    suspend fun createPdf(uiState: ConsentUiState): Boolean
+    suspend fun uploadPdf(pdfBytes: ByteArray): Boolean
 
 }
